[chore] #87 - 설명 없는 경우 spacing 조절

+++ Routee-iOS/Routee-iOS/Presentation/Common/Components/ActionPrimaryModal.swift
@@ -157,7 +157,7 @@ final class ActionPrimaryModal: UIViewController {
             contentStackView.setCustomSpacing(0, after: textStackView)
         } else {
             contentStackView.setCustomSpacing(
-                .s32,
+                .s16,
                 after: textStackView
             )
         }
